common: riscv64 port
As explained to me by Stefan O'Rear in 2018, cache control operations
have been intentionally excluded from the Unprivileged ISA, with the pmem
support design requiring CPU caches to be in the persistency domain.
This is analogous to eADR in Intel's implementation.

Thus, cacheline granularity is not supported, only byte and page.